requirements.txt
langchain
openai
google-search-results
apify-client
chromadb
tiktoken
chainlit
python-dotenv